combine all files into one
